Cost BOM Multi Level Review http://sourceforge.net/tracker2/?func=detail&atid=879335&aid=2504885&group_id=176962
Fix T_BOMLine.Cost reference type (changed from table dir to cost+price)

+++ base/src/org/eevolution/model/I_T_BOMLine.java
@@ -93,12 +93,12 @@ public interface I_T_BOMLine
 	/** Set Cost.
 	  * Cost information
 	  */
-	public void setCost (int Cost);
+	public void setCost (BigDecimal Cost);
 
 	/** Get Cost.
 	  * Cost information
 	  */
-	public int getCost();
+	public BigDecimal getCost();
 
     /** Column name CurrentCostPrice */
     public static final String COLUMNNAME_CurrentCostPrice = "CurrentCostPrice";
